pyproject.toml
[tool.poetry]
name = "automate-google-dinosaur"
version = "0.1.0"
description = ""
authors = ["Alen Vlahovljak <alen.vlahovljak@evolt.dev>"]
readme = "README.md"
packages = [{include = "automate_google_dinosaur"}]

[tool.poetry.dependencies]
python = "^3.9"
keyboard = "^0.13.5"
Pillow = "^9.2.0"
PyAutoGUI = "^0.9.53"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
